# web/openapi.json
{
  "components": {
    "schemas": {
...
      "BilibiliTaskSource": {
        "additionalProperties": false,
        "description": "A Bilibili source selected by URL.",
        "properties": {
          "kind": {
            "const": "bilibili",
            "title": "Kind",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "url"
        ],
        "title": "BilibiliTaskSource",
        "type": "object"
      },
      "BulkTaskMutationRequest": {
        "additionalProperties": false,
        "description": "One explicit operation over task identifiers in request order.",
        "properties": {
          "operation": {
            "enum": [
              "archive",
              "unarchive",
              "delete",
              "set_tags",
              "requeue"
            ],
            "title": "Operation",
            "type": "string"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "task_ids": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "minItems": 1,
            "title": "Task Ids",
            "type": "array"
          }
        },
        "required": [
          "operation",
          "task_ids"
        ],
        "title": "BulkTaskMutationRequest",
        "type": "object"
      },
      "BulkTaskMutationResponse": {
        "description": "Ordered outcomes for a bulk task mutation.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/BulkTaskMutationResult"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "title": "BulkTaskMutationResponse",
        "type": "object"
      },
...
      "CreateTaskRequest": {
        "properties": {
          "source_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "source_url"
        ],
        "title": "CreateTaskRequest",
        "type": "object"
      },
      "CreateWorkstationTaskRequest": {
        "additionalProperties": false,
        "description": "The v2 source, profile, and scheduling request for one new task.",
        "properties": {
          "priority": {
            "default": 0,
            "title": "Priority",
            "type": "integer"
          },
          "profile_id": {
            "const": "standard",
            "title": "Profile Id",
            "type": "string"
          },
          "source": {
            "discriminator": {
              "mapping": {
                "bilibili": "#/components/schemas/BilibiliTaskSource",
                "local": "#/components/schemas/LocalTaskSource"
              },
              "propertyName": "kind"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/BilibiliTaskSource"
              },
              {
                "$ref": "#/components/schemas/LocalTaskSource"
              }
            ],
            "title": "Source"
          }
        },
        "required": [
          "source",
          "profile_id"
        ],
        "title": "CreateWorkstationTaskRequest",
        "type": "object"
      },
      "CreateWorkstationTaskResponse": {
        "description": "The accepted v2 task identity and selected scheduling options.",
        "properties": {
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "profile_id": {
            "const": "standard",
            "title": "Profile Id",
            "type": "string"
          },
          "status": {
            "const": "pending",
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "profile_id",
          "priority",
          "status"
        ],
        "title": "CreateWorkstationTaskResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TaskListPage": {
        "description": "A page of task-library rows and its exact SQL count.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_count": {
            "title": "Page Count",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "page_count"
        ],
        "title": "TaskListPage",
        "type": "object"
      },
      "TaskOverview": {
        "description": "Complete workstation projection for one task.",
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "artifact_readiness": {
            "items": {
              "$ref": "#/components/schemas/ArtifactReadinessOverview"
            },
            "title": "Artifact Readiness",
            "type": "array"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/TaskArtifactOverview"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "execution_progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionProgressOverview"
              },
              {
                "type": "null"
              }
            ]
          },
          "pipeline_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Run Id"
          },
          "progress_percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Percent",
            "type": "integer"
          },
          "recovery_actions": {
            "items": {
              "discriminator": {
                "mapping": {
                  "download_asr_model": "#/components/schemas/DownloadAsrModelRecoveryAction",
                  "retry_stage": "#/components/schemas/RetryStageRecoveryAction"
                },
                "propertyName": "id"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/DownloadAsrModelRecoveryAction"
                },
                {
                  "$ref": "#/components/schemas/RetryStageRecoveryAction"
                }
              ]
            },
            "title": "Recovery Actions",
            "type": "array"
          },
          "safe_logs": {
            "items": {
              "$ref": "#/components/schemas/SafeLogOverview"
            },
            "title": "Safe Logs",
            "type": "array"
          },
          "source_kind": {
            "title": "Source Kind",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/TaskStageOverview"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "success",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "storage_bytes": {
            "minimum": 0.0,
            "title": "Storage Bytes",
            "type": "integer"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "title",
          "source_kind",
          "source_label",
          "status",
          "current_stage",
          "progress_percent",
          "tags",
          "storage_bytes",
          "created_at",
          "updated_at",
          "archived_at",
          "pipeline_run_id",
          "stages",
          "execution_progress",
          "artifact_readiness",
          "artifacts",
          "safe_logs",
          "recovery_actions"
        ],
        "title": "TaskOverview",
        "type": "object"
      },
      "TaskPatchRequest": {
        "additionalProperties": false,
        "description": "Editable task-library metadata, with absent fields left unchanged.",
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "maxLength": 80,
                  "minLength": 1,
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "TaskPatchRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Bilibili VTuber Suite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks": {
      "post": {
        "operationId": "create_task_endpoint_api_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Task Endpoint Api Tasks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task Endpoint",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "operationId": "task_detail_endpoint_api_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Task Detail Endpoint Api Tasks  Task Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Detail Endpoint",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/v2/tasks": {
      "get": {
        "description": "Return the requested bounded library page.",
        "operationId": "list_task_library_endpoint_api_v2_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "statuses",
            "required": false,
            "schema": {
              "items": {
                "enum": [
                  "pending",
                  "running",
                  "success",
                  "failed",
                  "cancelled"
                ],
                "type": "string"
              },
              "title": "Statuses",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "source_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Kind"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "updated_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updated From"
            }
          },
          {
            "in": "query",
            "name": "updated_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Updated To"
            }
          },
          {
            "in": "query",
            "name": "readiness",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ready",
                    "missing",
                    "failed",
                    "not_ready"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readiness"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated_at",
              "enum": [
                "updated_at",
                "created_at",
                "title",
                "storage_bytes"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Task Library Endpoint",
        "tags": [
          "workstation-tasks"
        ]
      },
      "post": {
        "description": "Create all legacy and workstation task records in one database transaction.",
        "operationId": "create_workstation_task_endpoint_api_v2_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkstationTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorkstationTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workstation Task Endpoint",
        "tags": [
          "workstation-tasks"
        ]
      }
    },
    "/api/v2/tasks/bulk": {
      "post": {
        "description": "Apply one lifecycle, tag, or requeue operation per requested task.",
        "operationId": "bulk_task_mutation_endpoint_api_v2_tasks_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkTaskMutationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkTaskMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Task Mutation Endpoint",
        "tags": [
          "workstation-tasks"
        ]
      }
    },
...
    "/api/v2/tasks/{task_id}": {
      "get": {
        "description": "Return one complete task-library overview.",
        "operationId": "task_library_overview_endpoint_api_v2_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Library Overview Endpoint",
        "tags": [
          "workstation-tasks"
        ]
      },
      "patch": {
        "description": "Replace the requested editable task metadata fields.",
        "operationId": "patch_task_library_endpoint_api_v2_tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Task Library Endpoint",
        "tags": [
          "workstation-tasks"
        ]
      }
    }
  }
}